refactor: separate sidebar and non-sidebar layouts into distinct route groups
- Split frontend layout into two route groups: (with-sidebar) and (without-sidebar)
- Move sidebar logic to dedicated sidebar-wrapper component
- Simplify root (frontend)/layout.tsx to minimal wrapper with global CSS
- Create clean layout for routes without sidebar (Header/Footer only)
- Fix React Client Manifest error by properly isolating client components

This prevents conflicts between sidebar and non-sidebar routes and ensures
proper component loading based on route group.
