+++ src/app/(frontend)/(with-sidebar)/flights/search/page.client.tsx
