+++ src/app/(frontend)/(without-sidebar)/(legal)/disclaimer/page.tsx
