+++ src/app/(frontend)/(without-sidebar)/(legal)/layout.tsx
