+++ src/app/(frontend)/(without-sidebar)/(legal)/privacy/page.tsx
