+++ src/app/(frontend)/(without-sidebar)/(legal)/terms/page.tsx
