+++ src/app/(frontend)/(without-sidebar)/auth/sign-up/page.tsx
