+++ src/app/(frontend)/(without-sidebar)/booking/confirmation/page.client.tsx
