+++ src/app/(frontend)/(without-sidebar)/booking/confirmation/queries.ts
