+++ src/app/(frontend)/(without-sidebar)/booking/payment/page.tsx
