+++ src/app/(frontend)/(without-sidebar)/home/info/page.tsx
