+++ src/app/(frontend)/(without-sidebar)/home/passengers/[id]/edit/page.tsx
