+++ src/app/(frontend)/(without-sidebar)/layout.tsx
@@ -1,0 +1,19 @@
+import { type ReactNode } from "react";
+
+import { Footer, Header } from "@/components/common";
+import { Toaster } from "@/components/ui/sonner";
+
+export default function WithoutSidebarLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return (
+    <div className="flex min-h-screen flex-col">
+      <Header />
+      <Toaster />
+      <main className="flex-1">{children}</main>
+      <Footer />
+    </div>
+  );
+}
